+++ app/Filament/Pages/Assessment/AstsHub.php
@@ -8,9 +8,11 @@ class AstsHub extends AssessmentTypeHubPage
 {
     protected static AssessmentType $assessmentType = AssessmentType::ASTS;
 
+    protected static string|\BackedEnum|null $navigationIcon = 'heroicon-o-document-check';
+
     protected static ?string $navigationLabel = 'ASTS';
 
     protected static ?string $slug = 'penilaian/asts';
 
-    protected static ?int $navigationSort = 2;
+    protected static ?int $navigationSort = 10;
 }